+++ playground/app/app.config.ts
@@ -1,0 +1,17 @@
+export default defineAppConfig({
+
+  ui: {
+    colors: {
+      primary: 'orange',
+      neutral: 'neutral'
+    },
+    icons: {
+      light: 'i-ph-sun',
+      dark: 'i-ph-moon'
+    },
+    skeleton: {
+      base: 'animate-pulse rounded-md bg-elevated'
+    }
+
+  },
+})