+++ src/runtime/components/PostComment.vue
@@ -1,13 +1,13 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 
-import { useNuxstrComments } from '../composables/useNuxstrComments'
+import { useComments } from '../composables/useComments'
 
 const props = defineProps<{ contentId?: string }>()
 
 const EMPTY_COMMENT = ''
 
-const { postComment } = useNuxstrComments(props.contentId)
+const { postComment } = useComments(props.contentId)
 const comment = ref(EMPTY_COMMENT)
 
 function isValidComment(commentText: string): boolean {
@@ -19,28 +19,38 @@ function clearComment(): void {
 }
 
 async function handlePost() {
+  if (!isValidComment(comment.value)) return
 
+  const wasPosted = await postComment(comment.value)
+  if (wasPosted) {
+    clearComment()
+  }
 }
 </script>
 
 <template>
-  <div class="text-sm text-muted-foreground border border-green mt-16">
-    <UTextarea
+  <div class="text-sm text-muted-foreground border border-green mt-4 p-6">
+    <div class="flex gap-2">
+      <div class="flex-1">
+
+      <UTextarea
       v-model="comment"
       class="w-full mb-4"
       placeholder="Write a comment ...."
       :rows="4"
     />
-    <div class="flex justify-end">
+      </div>
+    <div class="flex flex-col justify-center items-center p-2">
       <UButton
+        icon="mingcute:send-line"
         color="primary"
         variant="solid"
         :disabled="!comment.trim()"
-        class="mb-4"
+        class=""
         @click="handlePost"
-      >
-        Post Comment
-      </UButton>
+        size="xl"
+      />
+    </div>
     </div>
   </div>
 </template>